+++ sdk/storage/azure_storage_blob/tests/append_blob_client.rs
@@ -355,10 +355,9 @@ async fn test_create_append_blob_with_tags(ctx: TestContext) -> Result<(), Box<d
 
     let expected = HashMap::from([("env".to_string(), "test".to_string())]);
     append_blob_client
-        .create(Some(AppendBlobClientCreateOptions {
-            blob_tags_string: Some("env=test".to_string()),
-            ..Default::default()
-        }))
+        .create(Some(AppendBlobClientCreateOptions::default().with_tags(
+            HashMap::from([("env".to_string(), "test".to_string())]),
+        )))
         .await?;
 
     // Assert